pyproject.toml
[tool.poetry]
name = "quiz-app"
version = "0.1.0"
description = ""
authors = ["funashi55 <pinkgreen0304@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
langchain = "^0.3.9"
langgraph = "^0.2.53"
python-dotenv = "^1.0.1"
langchain-openai = "^0.2.10"
langchain-community = "^0.3.8"
langchain-core = "^0.3.21"
onnxruntime = "1.19.2"
langchain-chroma = "^0.1.4"
wikipedia = "^1.4.0"
fastapi = "^0.115.5"
firecrawl-py = "^1.6.1"
langgraph-checkpoint-postgres = "^2.0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
